// YiPet/src/definitions.d.ts
export interface ExtensionData {
    isEnabled: boolean;
    isReady: boolean;
    isAllowedFileSchemeAccess: boolean;
    settings: UserSettings;
    news: News[];
    shortcuts: Shortcuts;
    colorScheme: ParsedColorSchemeConfig;
    forcedScheme: 'dark' | 'light' | null;
    activeTab: TabInfo;
    uiHighlights: string[];
}

export type DevFixType = 'detector' | 'dynamic' | 'filter' | 'static';

export type DevToolsData = Record<DevFixType, string>;

export interface TabData {
    type: MessageTypeBGtoCS;
    data?: any;
}

export interface ExtensionActions {
    changeSettings(settings: Partial<UserSettings>): void;
    setTheme(theme: Partial<Theme>): void;
    setShortcut(command: string, shortcut: string): Promise<string | null>;
    toggleActiveTab(): void;
    loadConfig(options: {local: boolean}): void;
    applyDevFixes(type: DevFixType, text: string): Promise<void>;
    resetDevFixes(type: DevFixType): void;
    startActivation(email: string, key: string): void;
    resetActivation(): void;
    hideHighlights(ids: string[]): void;
}

export interface ExtWrapper {
    data: ExtensionData;
    actions: ExtensionActions;
}

export type ViewProps = ExtWrapper & {
    fonts?: string[];
};